Change funding rate to interest rate in projection
The "funding rate" so far was rather confusing. The issue with it was that we had multiple funding rates: one for going long and one for going short.
The fact that these two were different says that they are interest rates which need to be paid or are paid out for keeping a position open.

When looking at interest rates it also makes more sense to have negative or positive values and it's easier to see if you have to pay or will receive, i.e. similar to a bank account:
- If the interest rate is positive it means you receive,
- If the interest rate is negative it means you pay.

Combined with the position we can then convert the funding rate into interest rate:
As a reminder:
- A positive funding rate means "long pay short"
- A negative funding rate means "short pay long'