Store `PeerId` for `Cfd` in the database
In the model the `PeerId` is optional because we cannot always predict it.
For the maker it is impossible to map old taker identities to peer-ids.
For the taker we can only map known maker-ids to peer-ids.

In the database we use a `placeholder` `PeerId` to indicate if the peer-id of the counterparty is known.
This allows to add the `NOT NULL` constraint to the new `counterparty_libp2p_identity` column.
When loading the `counterparty_libp2p_identity` from the database for the taker, if the `placeholder` is loaded we default known maker peer-ids based on a mapping from the legacy network identity.
This allows new takers to use the new rollover protocol for positions that were opened with a previous taker version.

We introduce `HelloV3` message.
New takers connects to the maker with `HelloV3`.
The maker treats the taker's peer-id as optional field when passing on incoming messages; making the maker backwards-compatible.

Co-authored-by: Mariusz Klochowicz <mariusz@klochowicz.com>