Allow the tour to jump back to `/long` and re-render
This is a somewhat hacky solution to starting the tour (because of a refresh) from a different route.
Unfortunately we cannot detect the tour being started. The `onStart` callback is unfortunately only triggered upon enabling the tour, but not upon "starting" as in showing the first step.
The only possibility we have are the `onChange` functions - which all give us only the next element. Even `onBeforeChange` will display the current step.

This still works, but the second step will be displayed twice, once before we jump back to `/long` and once after (then in the right position, attached to the wallet button).