Depend on two versions of maia
With the advent of libp2p-based protocols, we have the opportunity to clearly
demarcate older and newer takers, making it possible to roll out recent maia bugfixes.

After this change:
- libp2p-based protocols will use maia v0.2.0
- legacy networking will use maia v0.1.0 (`maia_deprecated`), which matches maia
    version shipped with legacy takers

The contents of setup_contract.rs (protocol logic) has been copied wholesale to
`setup_contract_deprecated.rs` with the expectation that the whole file will be
deleted when we decide to drop support for takers using legacy networking layer.

Co-authored-by: Philipp Hoenisch <philipp@hoenisch.at>