Omit 0-amount outputs when reconstructing CETs
This fixes a bug where we were naively adding 0-amount outputs for the
party that was getting liquidated when reconstructing "liquidation"
CETs.

The network would not accept these transactions, but we were failing
as soon as we were checking against the original CET TXID. In `maia`
we were already constructing this kind of transaction correctly, by
omitting the 0-amount output.

Co-authored-by: Daniel Karzel <daniel.karzel@coblox.tech>