Fix issue where CFD is rejected and failed
This happened at an early stage when rejected and failed where not mutually exclusive yet.
In the current code base this should not happen.

Instead of using a specific CFD id to prevent the log we opt for a more generalized approach:
If the CFD was rejected it cannot be failed.