Verify validity of deconstructed CETs during debug builds
During contract setup and rollover, we deconstruct each individual CET
in order to save space. It is hard to write a test to verify that this
reconstruction logic is sound. Instead of writing a test we introduce
a `debug_assert_eq!` statement, which will only run on debug builds.

If we confirm that we can reconstruct a transaction with the original
TXID, we know that we did not mess up the deconstruction.