Reduce time to sync wallet
We only want to sync for 1000 addresses during the first sync. This is because we might have generated over 1000 addresses in the past and if we would not sync that far ahead we would miss some balance.
For repeatedly sync it is enough to sync for the default 100 addresses because everytime we generate a change address it is automatically included.