Split sync attestations and announcements
We only care about fetching new attestations after the initial load. Because, when a CFD is loaded from the DB, we add the needed `Attestation` (using `MonitorAttestation` msg) to a list which will then be fetched from the oracle on every `SyncAttestations` interval. Given initial loading from the DB takes some time, we start syncing only after the loading is complete.