Split xtra_libp2p_ping::Actor into two
Fixes https://github.com/itchysats/itchysats/issues/2050.

By splitting the original `xtra_libp2p_ping::Actor` into two simpler
actors we ensure that the handling of external messages (i.e. pongs
from the counterparty) is never halted by the processing of internal
messages.

Given how `xtra_libp2p::Endpoint` is designed, it is unnecessary (and
sometimes harmful) to combine the protocol roles of listener and
dialer in a single actor. Protocol listener actors such as `pong` are
registered with the `Endpoint` when it is constructed. Protocol
dialers (they don't need to be actors!) such as `ping::Actor` can
request a negotiated substream from the `Endpoint` to kick off (and
continue) the protocol.