Separate fetching connected peers from pinging
We don't want to block sending out all pings on synchronising our
internal state with the `Endpoint`. This could delay pings that the
listener is expecting.

We may end up failing to send pings to disconnected peers a bit more
frequently, but this was already possible before as you could not rely
on listeners still being connected by the time we sent out the ping:
the connection could have been dropped in between.