[NOMERGE] Verify that latency is relatively low
I have run variations of this test when working on
https://github.com/itchysats/itchysats/issues/2050.

The primary ways in which the test can be modified are:

- Number of `bob`s (equivalent to the `taker`s in `itchysats`).
- Whether the `bob`s send pings or not.
- Number of worker threads for the test.
- The ping interval (not investigated in depth).

The following observations assume that the fixes that appear in
earlier have been applied. Particularly, splitting the
`xtra_libp2p_ping::Actor` into two.

Some observations:

- If we don't let `bob`s send pings, we have <1ms latency with up to
10_000 bobs.

- If we do let `bob`s send pings, as we increase the number of `bob`s
we get more situations where we fall behind on particular
pings (although most of them are still returned in time). This
indicates to me that certain "pong" futures are not being scheduled,
leaving the "pinger" waiting for a very long time.

- Making the `tokio::test` multi-threaded and adding more and more
worker threads mitigates almost all problems. For example, we can run
the test for a minute with 200 pinging `bob`s with <1ms latency.