Stop takers from pinging the maker
Running a ping protocol between takers and makers serves several
purposes:

- Much like the heartbeat, using the connection regularly should help
keep it alive.
- The maker can gather metrics on the health of the connection between
itself and the takers that are connected to it.

Before this patch, each taker was also sending pings to the maker.
This was redundant (maker is already keeping the connection alive) and
wasteful (the takers were not doing anything with this information).

It was theorised that takers could eventually use their latency to the
maker to decide whether they should attempt to reconnect. I think
there are alternative ways to achieve this. For example, the taker
could trigger a reconnect if he has not received a ping from the maker
in a long time.