Introduce rollover version 3
Taker now default to version 3. The maker supports all versions.

This new version calculates the hours to extend the settlement time by
during rollover based on the event ID that will actually be proposed.
The motivation is to charge for rollovers based on the actual
extension to the settlement time of the CFD that they provide, fixing
https://github.com/itchysats/itchysats/issues/1974.