Do not cull all but one DLCs
The `cull_old_dlcs` API was meant to delete the old DLC data of all
CFDs. The implementation had a huge bug which instead deleted all DLC
data (old or current) for all open CFDs, except for the data of the
oldest DLC-related event ever.

We modify the test that was meant to verify that we don't delete
unexpected stuff when calling this API to make sure that we don't
introduce a similar bug in the future.

Fixes https://github.com/itchysats/itchysats/issues/2070.