Decouple legacy and libp2p socket initialisation
Allow network addresses on the actor system level - the networking stack can be
now configured separately.

Relying on some heuristics like `libp2p_port=legacy_port+1` inside the actor
systems caused problems in environments where we cannot ensure that both ports
next to each other are free.
In particular, this problem was observed when running the test harness in
parallel, resulting in "Address already in use" errors.

The heuristic above is still used in production, but contained in `main.rs`
files; it will be removed if all protocols will be using libp2p.