Constrain unbounded Tasks growth in rollover protocol
Instead of appending tasks indefinitely, constrain the rollover actor to only
allow one task per order ID at a time.
We implicitly rely on the invariants that we cannot get two decisions for the same rollover
at the same time.

Note: This is not a general solution, but it removes unbounded growth of the vector.