Let separate spawner::Actor manage pings
Fixes https://github.com/itchysats/itchysats/issues/2051.

It does so by spawning the ephemeral ping tasks via the
`xtras::spawner::Actor` (as opposed to using `Tasks` directly). This
actor will clean up the task handles internally when they are no
longer needed.

We are still able to tie the lifetime of theses tasks to the lifetime
of the `ping::Actor` because:

1. Actors are dropped when all their (strong) `xtra::Address`es are
dropped.
2. We store the only (strong) `xtra::Address` to the `spawner::Actor`
inside of `ping::Actor`.

This means that whenever `ping::Actor` is dropped, the last
`xtra::Address<spawner::Actor, Strong>` is dropped, stopping all the
ongoing tasks that were spawned through it.