Service not available instead of general HTTP error
Upon startup a `503` indicates in a better way that the service is just not available (yet).
This error can *only* happen on startup, because that is the only time the feed contains `None`.
We only initialize the feed with `None` but always send `Some`.

We could even go further and include a `Retry-After` header, but I did not implement that yet.