Make `now` explicit when calculating rollover hours
In the production code we make it more explicit that we are passing in `now`.
Being explicit can help us when changing this code - if we ever use `now` for derive another value in the same code-path we have to ensure to use the same now.

Test execution was flaky because we used different `now` in the same context.
So, when executing a test right before the full hour we might have diverging `now` values within a test resulting in unwanted behaviour.
In order to make to test this code deterministically we have to make now configurable.

We opt for introducing a function `common_time_boundaries` that returns an event at the full hour, one second after and one second before the full hour.
This should be sufficient.

Note that the proptest appears to have been wrong, we cannot do `+60` if we assume `now` to be the same.
We should not base this test on the fact that we have two different now values where one is slightly newer than the other one, for fixed values the boundary has to be `+59`.