No decision based on `peer_id` in maker rollover decision making
The maker cannot base his decision on the CFD's peer_id being set, because previously opened positions will not contain the peer_id and the maker cannot default it.
The idea was that the maker would use the `peer_id` known in the connection to decide, but accept/reject are completely independent of the connection.

To keep this logic simple and avoid passing on state to the cfd actor we just try to dispatch to libp2p and if that fails (because there is no pending rollover) we fallback to the legacy connection.
This removes the need of load the CFD which is nice.