Maker wallet address cache size configurable
This allows us to experiment with different cache sizes and overcome problems if a hardcoded address cache is not sufficient anymore.
Default is set to `500`. This is a change to the previously hardcoded cache of `1000` addresses. We use `1000` to be "on the safe side" for longer, but given that it is configurable now we don't have to be this exhaustive.