Use maia instead of maia_deprecated for libp2p-based collab settlement
Remove the test that was reproducing the maia_depracated bug, as it does not
panic anymore with new collab settlement protocol.

Note that the payout amounts in cfd tests had to be amended slightly, as the
way we calculate and split the fees based on the payout curve calculations is
different in maia and maia_deprecated.