Add `connection_monitor` actor
The `connection_monitor` regularly checks the endpoint and offsets the connection differences to define if a new connection was established or a connection was dropped.
This helps to overcome the problem that all actors interested in connection information bombard the endpoint with `GetConnectionStats`.
Furthermore, we can now push connection information to interested subscribers instead of polling this information.

The new `connection_monitor` actor is planned to be used by `ping` and other protocols.

Co-authored-by: Mariusz Klochowicz <mariusz@klochowicz.com>