Add notification capabilities to Endpoint
Furthermore, we can now push connection information (such as connected peer or
listen addresses) to interested subscribers instead of polling this information.

Roll out the usage in the existing actors, so they don't need to poll Endpoint continuously.

This helps to overcome the problem that all actors interested in connection
information bombard the endpoint with `GetConnectionStats`.

Co-authored-by: Mariusz Klochowicz <mariusz@klochowicz.com>