Use consistent naming in the settlement protocol
At some point we started using multiple terms for the same protocol:
"collaborative settlement", "settlement", "close position".

This commit removes the usage of the latter as it can be confused with
force-closing (according to bitMEX docs).

This is a separate commit in case we decide to revert the name change and go in
the other direction (ie. "close position").