Identify protocol
The `identify` protocol allows to request identity information from a peer.

Upon establishing a connection with a peer we request the identify information from the peer.

---

Add prometheus int gauge for libp2p connections

This replicates the legacy `NUM_P2P_CONNECTIONS_GAUGE` metric so we have similar stats available.

---

Maker protocol compatibility through `identify`

- Identify actor publishes `IdentifiedPeer`s in channel. Specifying the channel to be notified with peer information upon retrieval is optional because at the moment only the taker is interested in this information.
- The taker uses the `IdentifiedPeer` information to derive compatibility with the maker based on the protocols.

---

Display incompatibility warning upon incompatibility to maker

In addition to the `info` dialog that notifies the user that a new version is available the user will see a `warning` dialog in case the current version is incompatible with the connected maker. The incompatibility is based on the `unsupported_protocols` being defined. In the future we can take more elaborate decisions based on specific unsupported protocols if wanted.