Remove timestamp field from SettlementProposal
The motivation for this is that we don't want to compare the
timestamps generated by taker and maker when deciding if the maker
should accept a settlement proposal.

We could just modify the `PartialEq` implementation of
`SettlementProposal`, but it turns out that this field is never used.
We therefore remove it altogether.

We do no modify `wire::taker_to_maker::Settlement::Propose` because
this code is on the way out anyway. We simply ignore the `timestamp`
field if we receive it over the wire.