No refund-tx if CET was confirmed
It can happen that upon startup of the application that the refund timelock is already expired, but we publish the CET and not refund.
This is because the CET has priority over refund even if the refund timelock is already expired.
In `RefundTimelockExpired` we cannot know if the CET was confirmed or not (becuase this is a later state).
We opt for setting `refund_tx` to `None` and `refund_published` to `false` to ensure we do not wrongly show the refund-tx in the user interface if the CET was confirmed on chain.