xtras: Panic if send_interval fails
We rely on SendInterval trait to send periodic messages for the whole lifetime
of an actor. If for whatever reason we stop sending periodic messages, we should
panic to trigger spawning a new instance. Otherwise, we risk at continuing
running the actor that *appears* to work on the surface level, but in reality
some of its functionality is crippled.