Don't sort events by timestamp
It appears that it can happen that two events in the database have the same timestamp.
This was observed for collab started and completed as well as rollover started and accepted events.
If they have the same timestamp, and the event history is sorted by that timestamp it can happen that the order is incorrect.

Iterating over the list of events should not change any ordering.
Thus, the database query can ensure the order and we don't need an additional sorting step.