Ensure message order at the end of rollover
This should mitigate a problem related to libp2p-rollover, where the taker closed a substream too early.
If both parties send and then wait for the other party in parallel it is unclear who will receive first.
It can happen that the taker sent and received and closed the stream, while the maker is still reading the message that was sent by the taker. This can lead to a panic because the substream is already close.
By ensuring that the taker first sends and then waits to receive, and the maker first waits to receive and then sends should solve this problem.

This is backwards compatible, because old takers will do the same as before (for send then wait to receive).