Don't add PRs to project board
There is not much value in having every PR in the project board. Instead, we should link PRs to tickets they are solving. If there is no ticket, you can always manually add that PR to the board. There is a limit of max items in the project board of 1200 which we have reached. Most of them are version bumps from dependabot