Poll for connected peers when dialer connects
If a dialer is restarted, it might lose the information whether the endpoint is
connected; in other words, we cannot rely on the subscribed information
unless we build current information at startup.

Simplify the implementation by removing `tasks` and `connected` state; we now
try to connect, and if it doesn't work, then we shut down the actor.
Note that it is not necessary to subscribe for `ConnectionEstablished` message
anymore, as we poll for that information once to verify that request has worked.