Double rollover test
Testing one rollover is good, but we are asserting on `Open` in the end, we are not actually asserting the rollover worked.
We sometimes have cases where the first rollover succeeds but then consecutive rollovers fail.
By rolling over twice we have more assurance.

We keep the single rollover tests so we can more easily know which part failed when running the tests.
If the single test fails we know that the problem must be with the first rollover, if only double failed we know it is the consecutive rollover.