Ensure rollover actually works by asserting on fees
Previously we only asserted on the state being `Open` again, but even if the rollover failes we would be in an `Open` state.
To ensure the rollover actually worked the easiest way is to assert that the fees changed (because one open the fees only change with a rollover).

The funding rate of our test offer is always positive, i.e. long pays short.
If one is long and long pays short then the fees should be more, if one is short and long pays short the fees should be less after a rollover.

We don't assert on the exact fee because it is good enough to know that it is higher/lower than before.