Add restart timeouts to dialer and listener actors
In an event of network problem, these actors will try to respawn immediately.
This is counterproductive, because it will only incur an unnecessary load on the
application, with very little chance of succeeding.

If dialing/listening fails, don't try to restart immediately. Instead, wait
specified time before the next attempt.