Introduce /itchysats/order protocol over libp2p
- The key test `taker_takes_order_and_maker_accepts_and_contract_setup`
passes. Adapting the rest of the actor tests should be trivial.

- I am not sold at all on using an internal channel over handling the
decision at the actor level on the maker. The current design was
motivated by the possibility of testing the protocol in isolation, but
it's basically impossible given all the coupling with actors in
`daemon`.

- The file `taker_cfd.rs` is not very clean. Conversely,
`maker_cfd.rs` gets more complicated because we have to support the
legacy connection for a while longer. The logic there is to first try
to accept/reject using the new protocol and if it fails attempt it
over the legacy connection. This is very stupid, but I would rather do
something stupid rather than trying to differentiate between different
types of CFDs.

- You can see that the new design follows the proposal in
https://github.com/itchysats/itchysats/issues/1770. This means that
the taker now generates an `OrderId` when _placing_ the order. We
currently reuse the same type for the offer (legacy) and the taker's
order, which is a bit confusing (but hard to avoid, because our model
wants an `OrderId`). With the new protocol the taker ignores the
`OrderId` associated with the offer, but we have to keep it around for
legacy connections.