Save `settlement_event_id` with `RevokeCommit`
This is necessary so the maker can make decisions if a rollover can be started from a specific commit.
The maker has to validate that the event-id associated with the commit is not in the past yet, otherwise a rollover is not allowed anymore.

For backwards compatibility the `settlement_event_id` in `revoke_commit_transactions` is an optional field.
One all legacy cfds are closed we could add a `NOT NULL` constraint and remove the `Option` from the business logic.