Use `CompleteFee` for rollover
Instead of accumulating the fees we apply the complete fee at the end of each rollover in `RolloverCompleted`.
This is necessary, because the maker now allows multiple rollovers for the same `event_id` resulting in problems when accumulating the `funding_fee`s when applying `RolloverCompleted` events.
By saving the complete fee we introduce a snapshot at that point in time.

To ensure backwards compatibility we make the `complete_fee` in `RolloverCompleted` is optional.
If present we create the `FeeAccount` from the `complete_fee` otherwise we use the old logic of accumulating using the `funding_fee` of the rollover.

Additionally, we introduce the `CompleteFee` in the `RevokeCommit` to provide a fee-snapshot for previous rollovers.
This enables the maker to accept rollover requests for a previous `commit_txid`.
To ensure backwards compatibility we make the `complete_fee` in `RevokeCommit` is optional.