Fix redirection bug
All the useLatestEvent, useWebsocket, etc - hooks make the `App` load more than once.
Because of this, the website is rendered repeatedly.
Given the default url `/` had a redirect to `/long` we ended up redirection repeatedly until the browser stopped us. The result was, an empty page on the first load.
To overcome this I added a manual logic which only runs on a page refresh, i.e. if you change the url or land on the side the first time it will be triggered.
The result is a faster loading and less redirection.
The downside is, that we will need to keep the logic up-2-date if we ever want to add a new route.