Maker allows rollover from `commit_txid`
The taker sends a `commit_txid` from where to roll over in the rollover `Proposal`.
Based on this `commit_txid` the maker picks the `event_id` from which to roll over. The costs are determined according to that `event_id`.
The maker does not allow to roll over from event id that has likely occurred.

This allows us to retry rollovers as the taker.
If the maker is ahead by one rollover, the taker can trigger a rollover with the previous event id and the maker will roll over from that point.