Maker fires and forgets last message
In the new protocol the maker ensures to finish the protocol and append a `RolloverCompleted` if the maker received the revocation keys from the taker.
If the maker fails to send the revocation keys to the taker (or the taker fails to receive) the expectation is that the taker will fail the rollover and retry using the same `event-id` as before.

This is safe for the maker because it ensures to complete (the taker cannot attack the maker by closing the connection).
This is resilient for maker and taker because the taker can retry there is a failure.