Introduce /itchysats/order protocol over libp2p
The new design follows the proposal in #1770. This means that
the taker now generates an `OrderId` when _placing_ the order. We
currently reuse the same type for the offer (legacy) and the taker's
order, which is a bit confusing (but hard to avoid, because our model
wants an `OrderId`). With the new protocol the taker ignores the
`OrderId` associated with the offer, but we have to keep it around for
legacy connections.

One actor test got removed - with libp2p we're not changing the offers after
one was picked; the OrderId is generated on the taker side.

Co-authored-by: Lucas Soriano del Pino <l.soriano.del.pino@gmail.com>