Introduce /itchysats/order protocol over libp2p
When placing an order from the taker to the maker the taker sends the `OfferId` that the order is for.
The taker generates an `OrderId` when _placing_ the order.
The maker evaluates if the order is valid based on the `OfferId` being still the current one.
This allows one/multiple takers to take the same offer, resulting in different orders.
The `offer_id` is saved as additional id in the `Cfd`.

We currently reuse the same type for the offer (legacy) and the taker's order, which is a bit confusing (but hard to avoid, because our model wants an `OrderId`).

With this change the `Order` depends on the `Offer`, but we can handle them individually in data.
Since we are currently not interested in storing the `Order` data yet, we don't store the `Order` separately but keep the `Cfd` data flattened out.
The `Offer` is not stored.

Co-authored-by: Lucas Soriano del Pino <l.soriano.del.pino@gmail.com>
Co-authored-by: Mariusz Klochowicz <mariusz@klochowicz.com>