Avoid endlessly re-connecting in case a connection was already established
This can happen in scenarios where the connection is still being established, but the dialer ran into a timeout already.
The dialer then restarts, but when asking the endpoint for connection information it is connected at that point - it then stops and restarts endlessly (until we lose connection which could be never).