Instrument routes, endpoint, sleep, and timeout
and errors. The spans are named `[method] [route]`, e.g `GET /feed`.
Add `#[instrument]` macros to all rocket routes, logging their arguments, returns,
and errors. The endpoint and dialer actor were also instrumented.

This also adds instrumented `sleep` and `timeout` functions, which wrap the tokio equivalents
with more detailed spans and logs. The `timeout` function also requires a child span
for the future that is beign timed out, in order to prevent accidentally forgetting
to instrument the child span, and to make sure the child span is a child and not
sibling of the "Future with timeout" span