Extract rollover protocol into separate crate
The main motivation for this change is to be able to support multiple
versions of the rollover protocol at the same time, without incurring
in lots of duplicated code or increased code complexity due to
branching logic.

Additionally, I think this should allow us to test this protocol in
isolation. I think it's also worth it in terms of the overall
structure of the workspace, and could lead to improved compilation
times.

In my view, the main downside of this approach is that the API of the
rollover protocol is slightly more complicated. We can no longer rely
on knowing the actors and components that we interact with, because
they remain in `daemon`. Instead we depend on 2 new traits:
`model::ExecuteOnCfd` and `rollover::GetAnnouncements`.

The upside of making the rollover actors generic via these traits is
that we can now test this protocol in isolation, as mentioned above.