Log timeouts inside of handler span and set interrupted
This sets interrupted=timed_out on handler spans that time out, which could be helpful for filtering spans by which time out, instead of having to rely on just looking for spans that take longer than others. Furthermore, the "Handler execution timed out" log message will be attached to the handler span. Xtra was updated to facilitate this.