`ConnectionStatus` is owned by `online_status` actor
We should not depend on legacy connection types, so the `ConnectionStatus` was moved to the `online_status` actor because it is part of the API of that actor.
While moving it, the `reason` in `Offline` was removed because it is not used by the `online_status` actor.
The user interface will use the `maker_identity` instead to determine compatibility.