Display incompatibility warning upon incompatibility to maker
In addition to the `info` dialog that notifies the user that a new version is available the user will see a `warning` dialog in case the current version is incompatible with the connected maker. The incompatibility is based on the `unsupported_protocols` being defined. In the future we can take more elaborate decisions based on specific unsupported protocols if wanted.