Ensure to monitor for finality after `LockConfirmedAfterFinality`
We should not treat `LockConfirmedAfterFinality` as final state in the monitor, because this would mean we just stop monitoring for other things.
This can lead to scenarios where we never confirm collab settlement or CET in cases where `LockConfirmedAfterFinality` is recorded before the finality confirmation is picked up.

Treating `LockConfirmedAfterFinality` like `LockConfirmed` is correct and should not have side effects in the monitor.
`LockConfirmedAfterFinality` should not have any influence on other monitoring decisions besides the lock transaction.