Update rollover protocol to use liquidation events
The new `itchysats/rollover/2.0.0` protocol adds liquidation events to
the DLC.

In order to preserve backwards compatibility with takers running
`itchysats/rollover/1.0.0` we run another rollover actor in parallel
for the maker. This leads to a ton of duplicated code, but this
technical debt is temporary and contained in the
`xtra-libp2p-rollover` crate. The idea is to release all our crates to
`crates.io`, so that we can eventually introduce a single cargo
dependency per version of a protocol.