Prepare setup_contract::new to use liquidation events
This code used to assume that we had a single event per CFD when
constructing the corresponding DLC. After introducing liquidation
events, that is no longer the case.

Caveat: this code is currently not used! When we added patch
99e62ba6c2d046ec848878c73a15ed2ffed13a40, we chose to only use the
deprecated version of this code (in `setup_contract_deprecated.rs`)
for the maker and all possible takers. The idea was to then use
`setup_contract::new` after adding the contract setup protocol over
the new network layer. Since this hasn't happened at the time of
writing this message, we cannot test the changes in this patch yet.