Do not block Endpoint when opening substreams
It is sufficient to just look for the corresponding `yamux::Control`
for a particular peer inside `Endpoint` and then release the lock (by
returning from the handler).

The actual opening of the substream using `yamux` can be delegated to
a future which is returned to the caller, who must `await` it.