Refactor send_async_safe
Send_async_safe for Result-returning messages now is a bit more traceable, as it will first send, resolving to the receiver, and then spawn a task awaiting the receiver and printing the error if there is one. This is more traceable, as the send span will be shown under the calling span.