Use send_async_next when we dispatch an internal task to finish processing
In many places where we use `send_async_safe` we want to dispatch a follow-up
task that is still relevant to the current request we're handling, but we do
not want to block the message handler. It is convenient to schedule such tasks
as messages, but in order to guarantee they would not be subjected to
backpressure or queued up behind subsequent requests (as we want to action them
when their context is still relevant), we need to be able to schedule them with
higher priority than regular messages arriving on our inbox.

Concrete examples from this changeset:
- Endpoint wants to process newly created/dropped connection as soon as it can,
instead of scheduling it on the queue.
- updating metrics after a handler