Spawn wallet actor as a blocking task
The wallet actor performs a lot of CPU-bound, blocking work in its handlers. This is
bad, as it currently runs on the tokio runtime, which explicitly forbids blocking like
this. It also does not have any asynchronous handlers, so this commit makes the wallet
actor run on a separate thread which blocks on the execution of the actor's future.