Ignore frequent spans
Using a Directive in an EnvFilter didn't seem to do anything at all so I landed on this. This was very painful as reordering anything at all in the registry init lead to completely broken output. It is still not perfect as it will ignore every span, even if it has an error.