Build new RevokedCommit based on commit TXID proposed by taker
The taker already did so because they could simply _always_ take the
information from the latest DLC.

The maker was supposed to do so, but they were naively using the
information from the latest DLC. They were basically ignoring the
`from_commit_txid` field in the message as soon as they had confirmed
that they could find it in their commit transaction records.

We also had to fix the rollover retry test, which used to pass even
though this bug existed:

On `testnet`, when the maker is ahead with rollovers we run into error
`Counterparty sent invalid revocation sk` when rolling over from a
previous commit. This was not detected by the
`retry_rollover_an_open_cfd` test because, even though we trigger
rolling over from a previous commit-txid, the taker's DLC actually
matched the one of the maker. The test setup did no properly simulate
that the maker is in fact one rollover ahead of the taker.

As it is hard to remove events, the easiest way to simulate the
previous rollover on the taker side, is to add another
`RolloverCompleted` event with the old data. This should be sufficient
to simulate that the taker has in fact the old state and the maker is
ahead.

Co-authored-by: Daniel Karzel <daniel.karzel@coblox.tech>