Make tmp db file absolute
It seems that relative path-based execution leads to errors from time to time where the db file can't be found. Using an absolute path solves this problem. The script needs to be executed from within sqlite-db module anyway to produce the sqlx-data.json file. Hence, I think this is a justified solution.