Allow authenticated ws to BitMEX
Some APIs of BitMEX expect authentication such as retrieving updated for `Order`, `Executions` or `Position`.
By passing in `credentials` the library will try to send an authentication message.

I've also changed the way how we subscribe to topics: instead of passing them into the url, a explicit subscription method was added.

Note: if authentication fails, an error message as string will be returned and if a topic was chosen which requires authentication the stream will be closed.