WIP - Link offer-id to order-id
Instead of the taker sending all parameters to the maker when taking an offer, the taker adheres to the offer, i.e. the order is associated with an offer-id so the maker know what the taker refers to and does not have to validate all the values.
Both parties then start the contract setup according to the values of the offer + the values decided by the taker (