Ignore digits when deriving BitMexPriceEventId's PartialOrd
This is important because deriving `PartialOrd` normally means that it
looks at all the struct's fields _in order_. For this type we want to
compare the `timestamp`, but do not care about the `digits`! More
importantly, someone could unintentionally change this behaviour by
reordering the fields.

We also have to ignore the `digits` field for other derived traits,
because inconsistencies there could cause trouble[1].

[1]: https://github.com/rust-lang/rust-clippy/issues/1621.

Co-authored-by: Daniel Karzel <daniel.karzel@coblox.tech>