Tell users to include contact details in details
At the moment it is not obvious that we cannot contact users when they send a request via feedback fish.
There is no means to specify additional fields in the feedback fish form.

We can opt for asking users for their emails to "register" their ItchySats version, and then attach the email as metadata, but this is a separate feature that will require some more time.
The easiest way to tell the user that they have to provide contact details in order to receive a response is a tooltip.