Remove support for legacy rollover protocol
The original motivation for this patch was to simplify the codebase by
removing any code related to the maker's manual accept/reject rollover
mechanism, which was removed from the `/itchysats/rollover` protocol
with https://github.com/itchysats/itchysats/pull/2533.

But doing so would mean that we would have to update the legacy
rollover protocol on the maker side (not a breaking change since it
only affects the maker, but a change I don't want to make). I think we
can already stop maintaining this code if we intend to stop support
for the legacy rollover protocol in the next release.