Taker drops connection upon failure to open substream for rollover
Rollovers are important to keep the position alive.
If the taker fails to open a substream for rollover then something might be wrong with the connection.
To make sure we don't run into a scenario where the actor cannot open substreams the taker drops the connection.

This is an emergency fallback scenario that should usually not be triggered.
It is added to avoid force closing positions due to connection problems.