Simplify open_cfd test helper
The `FeeCalculator` is not used in `open_cfd` and can be built before
calling it.

We also verify that when opening a CFD we choose a `taker_leverage`
which is actually supported by the `leverage_choices` configured in
the `OfferParams`. And that we use the expected `taker_leverage` when
calling `place_order`!