Set rustfmt max comment and line width to 100
It was a bit weird that the `comment_width` was configured to be
_longer_ than the `max_width` which sets the maximum width of each
line.

I choose to explictly set `max_width` to its default value of 100 for
clarity.

The reason why we don't see any other changes to the diff should be
obvious: setting `comment_width` to more than `max_width` was useless!

Inspired by:
https://github.com/rust-lang/rustfmt/issues/3349#issuecomment-622522913.