Introduce currency selector
We have to do some not-so-nice hacks to get into the location from App.tsx because we are not in a sub-route yet. Ideally, only the subroute would need to know about what symbol currently is in the path.

I stopped here with refactoring, but I think this can be cleaned up and things can move a few levels down.