Document payout curve bug
Payout curve price intervals are expected to be monotonically
increasing. That means that they should never go backwards. If they
did, they would overlap with previous intervals. That would cause more
than one CET to be activated when the oracle attests to a price in the
overlapping intervals.

In fact, if this happens we would fail when calling
`maia_core::generate_payouts` during contract setup or rollover.

We add a proptest that fails, but mark it with the `ignore` attribute.
The bug seems to occur depending on the `initial_price`,
`long_leverage`, `short_leverage` and `n_payouts`. The other arguments
appear to have no effect on whether we run into this bug or not.