Prove that OraclePayouts can be built from Payouts
The scope of the proptest added is limited by the bug documented in
the parent commit.

Nonetheless, if we choose a set of values that don't run into the
aforementioned bug (unrelated to the code we're testing here), we
demonstrate that the property holds: that we can build `OraclePayouts`
from `Payouts` with the expected number of events and the expected
number of liquidation payouts (2 per event).